Created a new `migrate` command
This commit introduces Goose as the migration tool. Using goose as a module, storeman can embed migration files into the binary.

Unfortunately, the only place I can think of the make this work is within the `/cmd/migrate` directory. This is un-intuitive and feels like a code smell. If I can find a better solution in the future - one that has the migrations at the root of this project, I'll use that.

Signed-off-by: Daniel Michaels <dan@danielms.site>