Add data layer (#3)
* Add sqlite and goose to dependencies

* Store and sqlite data layer

The data layer using Sqlite as the database created. Store added to Server struct and DB connection pool opened when the webserver starts.

Signed-off-by: Daniel Michaels <dan@danielms.site>

* Created a new `migrate` command

This commit introduces Goose as the migration tool. Using goose as a module, storeman can embed migration files into the binary.

Unfortunately, the only place I can think of the make this work is within the `/cmd/migrate` directory. This is un-intuitive and feels like a code smell. If I can find a better solution in the future - one that has the migrations at the root of this project, I'll use that.

Signed-off-by: Daniel Michaels <dan@danielms.site>

* explicitly add CGO_ENABLED=1 to `.air.toml` as go-sqlite3 needs it.

Signed-off-by: Daniel Michaels <dan@danielms.site>

* Added misc comments

Signed-off-by: Daniel Michaels <dan@danielms.site>

* fix -vet errors

Signed-off-by: Daniel Michaels <dan@danielms.site>